Add success and error states to the button
